New GltfMixin, which GltfCatalogItem now uses
